# Reducing latency for constructing and transmitting a response