--- a/fastwsgi/request.c
+++ b/fastwsgi/request.c
@@ -24,9 +24,9 @@ void logrepr(int level, PyObject* obj) {
     Py_XDECREF(str);
 }
 
-static void set_header(PyObject* headers, const char* key, const char* value, size_t length) {
+static void set_header(PyObject* headers, const char* key, const char* value, ssize_t length) {
     LOGi("setting header");
-    int vlen = (length > 0) ? (int)length : (int)strlen(value);
+    ssize_t vlen = (length >= 0) ? length : strlen(value);
     PyObject* item = PyUnicode_FromStringAndSize(value, vlen);
 
     PyObject* existing_item = PyDict_GetItemString(headers, key);
@@ -527,10 +527,10 @@ void build_wsgi_environ(llhttp_t* parser) {
     PyObject * headers = client->request.headers;
 
     const char* method = llhttp_method_name(parser->method);
-    set_header(headers, "REQUEST_METHOD", method, 0);
+    set_header(headers, "REQUEST_METHOD", method, -1);
     const char* protocol = parser->http_minor == 1 ? "HTTP/1.1" : "HTTP/1.0";
-    set_header(headers, "SERVER_PROTOCOL", protocol, 0);
-    set_header(headers, "REMOTE_ADDR", client->remote_addr, 0);
+    set_header(headers, "SERVER_PROTOCOL", protocol, -1);
+    set_header(headers, "REMOTE_ADDR", client->remote_addr, -1);
 }
 
 void init_request_dict() {