--- a/fastwsgi/request.h
+++ b/fastwsgi/request.h
@@ -16,8 +16,7 @@ typedef enum {
     RF_HEADERS_PYLIST  = 0x02
 } response_flag_t;
 
-int build_response_ex(void * client, int flags, int status, const void * headers, const char * body, int body_size);
-void build_response(PyObject* wsgi_response, StartResponse* response, llhttp_t* parser);
+int build_response_ex(void * client, int flags, int status, const void * headers, const void * body, int body_size);
 
 llhttp_settings_t parser_settings;
 void configure_parser_settings();