Before changes:
```
$ wrk -t1 -c1 -d30 http://172.16.220.205:5000/ --latency
Running 30s test @ http://172.16.220.205:5000/
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    17.39ms  496.59us  22.31ms   69.79%
    Req/Sec    56.93      4.62    60.00     69.33%
  Latency Distribution
     50%   17.44ms
     75%   17.71ms
     90%   17.96ms
     99%   18.43ms
  1708 requests in 30.01s, 13.01GB read
Requests/sec:     56.91
Transfer/sec:    443.75MB
```

After changes:
```
$ wrk -t1 -c1 -d30 http://172.16.220.205:5000/ --latency
Running 30s test @ http://172.16.220.205:5000/
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    11.23ms  717.36us  22.40ms   80.60%
    Req/Sec    87.80      4.69   101.00     73.33%
  Latency Distribution
     50%   11.26ms
     75%   11.64ms
     90%   11.89ms
     99%   12.28ms
  2634 requests in 30.01s, 20.06GB read
Requests/sec:     87.78
Transfer/sec:    684.64MB
```

Payload: 8MB JSON (from Flask)