@jamesroberts 2022-10-29
Thanks for this, this does seem to drop the latency a bit. However, there is a significant performance drop (almost 5x RPS) after PR #19. Could you take have a look into it?